scripts/requirements.txt
pandas
gspread
google-auth
psycopg2-binary
python-dotenv
